src/main/java/com/crud/tasks/trello/facade/TrelloFacade.java
package com.crud.tasks.trello.facade;

import com.crud.tasks.domain.*;
import com.crud.tasks.mapper.TrelloMapper;
import com.crud.tasks.service.TrelloService;
import com.crud.tasks.trello.validator.TrelloValidator;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.stereotype.Component;

import java.util.List;
import java.util.stream.Collectors;

@Component
@RequiredArgsConstructor
@Slf4j
public class TrelloFacade {
    private final TrelloService trelloService;
    private final TrelloMapper trelloMapper;
    private final TrelloValidator trelloValidator;

    public List<TrelloBoardDto> fetchTrelloBoards() {
        List<TrelloBoard> trelloBoards =  trelloMapper.mapToTrelloBoard(trelloService.fetchTrelloBoards());
        List<TrelloBoard> filteredBoardList = trelloValidator.validateTrelloBoards(trelloBoards);

        return trelloMapper.mapToTrelloBoardsDto(filteredBoardList);
    }

    public CreatedTrelloCardDto createTrelloCard(final TrelloCardDto trelloCardDto) {
        TrelloCard trelloCard = trelloMapper.mapToCard(trelloCardDto);
        trelloValidator.validateCard(trelloCard);

         return trelloService.createTrelloCard(trelloMapper.mapToCardDto(trelloCard));
    }
}
